FEBioStudio: removed duplicate CMathParser class. Removed #ifdefs in header files in favor of pragma once declarations.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBioStudio@12475 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ MeshLib/FEFaceEdgeList.cpp
@@ -400,7 +400,7 @@ FEEdgeFaceList::FEEdgeFaceList(const FEMesh& mesh)
 	{
 		const FEFace& face = mesh.Face(i);
 
-		int ne = FEL[i].size();
+		int ne = (int) FEL[i].size();
 		for (int j=0; j<ne; ++j)
 		{
 			int iedge = FEL[i][j];
@@ -427,7 +427,7 @@ FEEdgeFaceList::FEEdgeFaceList(const FESurfaceMesh& mesh)
 	{
 		const FEFace& face = mesh.Face(i);
 
-		int ne = FEL[i].size();
+		int ne = (int) FEL[i].size();
 		for (int j = 0; j<ne; ++j)
 		{
 			int iedge = FEL[i][j];
